requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
alpaca-py==0.31.0
alpaca-trade-api==3.2.0
annotated-types==0.7.0
attrs==25.1.0
certifi==2024.8.30
charset-normalizer==3.4.0
deprecation==2.1.0
frozenlist==1.5.0
idna==3.10
matplotlib==3.7.1
msgpack==1.0.3
multidict==6.1.0
numpy==2.2.6
packaging==24.2
pandas==2.2.3
propcache==0.2.1
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.1
requests==2.32.3
scipy==1.13.1
six==1.16.0
sseclient-py==1.8.0
tabulate==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
websocket-client==1.8.0
websockets==10.4
yarl==1.18.3